requirements.txt
Django==4.2.7
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.15.0
pandas==2.2.3
celery==5.3.4
redis==5.0.1
python-decouple==3.8
webdriver-manager==4.0.1
google-api-python-client>=2.70.0
google-auth>=2.20.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=1.0.0